Make DEBUG() use note() instead of say()
This is not only for semantics (debug info is more appropriate to
standard error than standard out), but also as a practicality; people
may want to enable debug output, but pipe that into another file. This
allows you to do that while printing standard out messages to the
terminal (or wherever that stream is redirected).